+++ common/src/main/java/io/github/mortuusars/envelope/world/entity/ai/goal/PigeonWanderGoal.java
@@ -23,7 +23,9 @@ public class PigeonWanderGoal extends Goal {
 
     @Override
     public boolean canUse() {
-        return !pigeon.isSitting() && pigeon.getNavigation().isDone() && pigeon.getRandom().nextInt(25) == 0;
+        return !pigeon.isSitting() && pigeon.getNavigation().isDone()
+              && (pigeon.level().getGameTime() - pigeon.getPigeonholeHandler().getLastTickInside() < 20
+                || pigeon.getRandom().nextInt(25) == 0);
     }
 
     @Override