+++ neoforge/src/main/java/io/github/mortuusars/envelope/neoforge/datagen/DataGeneration.java
@@ -12,6 +12,7 @@ import net.neoforged.neoforge.common.data.DatapackBuiltinEntriesProvider;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
 import net.neoforged.neoforge.data.event.GatherDataEvent;
 
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 @EventBusSubscriber(modid = Envelope.ID)
@@ -30,6 +31,9 @@ public class DataGeneration {
         generator.addProvider(event.includeServer(), blockTags);
         generator.addProvider(event.includeServer(), new ItemTagsDatagen(output, registries, blockTags.contentsGetter(), existingFileHelper));
         generator.addProvider(event.includeServer(), LootTablesDatagen.create(output, registries));
+        generator.addProvider(event.includeServer(), new AdvancementsDatagen(output, registries, existingFileHelper, List.of(
+              new AdvancementsDatagen.Generator()
+        )));
 
         DatapackBuiltinEntriesProvider datapackRegistries = new BuiltInDatapackEntries(output, registries);
         generator.addProvider(event.includeServer(), datapackRegistries);