+++ neoforge/src/main/java/io/github/mortuusars/envelope/neoforge/datagen/server/loot/GameplayLootProvider.java
@@ -1,0 +1,27 @@
+package io.github.mortuusars.envelope.neoforge.datagen.server.loot;
+
+import io.github.mortuusars.envelope.Envelope;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.loot.LootTableSubProvider;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.storage.loot.LootPool;
+import net.minecraft.world.level.storage.loot.LootTable;
+import net.minecraft.world.level.storage.loot.entries.LootItem;
+import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
+
+import java.util.function.BiConsumer;
+
+public record GameplayLootProvider(HolderLookup.Provider registries) implements LootTableSubProvider {
+    @Override
+    public void generate(BiConsumer<ResourceKey<LootTable>, LootTable.Builder> output) {
+        output.accept(
+              Envelope.LootTables.PIGEONHOLE_WASTE,
+              LootTable.lootTable().withPool(LootPool.lootPool()
+                    .setRolls(ConstantValue.exactly(1))
+                    .add(LootItem.lootTableItem(Items.BONE_MEAL).setWeight(100))
+                    .add(LootItem.lootTableItem(Items.DIAMOND).setWeight(1))
+              )
+        );
+    }
+}